+++ cracker-client/src/+utils/routes.js
@@ -2,6 +2,5 @@ export const ROUTES = {
   CALLBACK: "/callback",
   HOME: "/",
   LOGOUT: "/logout",
-  MAP: "/map",
   MARKERS: "/markers",
 };
