+++ cracker-client/src/markers/MarkersView.jsx
@@ -1,17 +1,19 @@
 import React, { useState } from "react";
+import auth from "../+utils/Auth";
 import { Row, Col } from "antd";
-import { MapView } from "../map/MapView";
+import { MapView } from "./map/MapView";
 import { Description } from "./description/Description";
 import { MarkerContext } from "./+context/MarkerContext";
 
 export const MarkersView = () => {
+  const isAllowed = auth.isAuthenticated() && auth.isUserAdmin();
   const [currentMarker, setCurrentMarker] = useState(null);
 
   return (
     <MarkerContext.Provider value={{ currentMarker, setCurrentMarker }}>
       <Row>
         <Col span={12}>
-          <MapView />
+          <MapView isAllowed={isAllowed} />
         </Col>
         <Col span={12}>
           <Description />
