+++ cracker-client/src/markers/map/+hooks/useRemoveMarker.jsx
